viewer/pyproject.toml
[tool.poetry]
name = "uwbins-viewer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
rerun-sdk = "0.21.0"
pandas = "^2.2.3"
numpy = "1.26.2"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
